requirements.txt
openai==0.27.0
python-pptx==0.6.19
